# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CannaZen e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, rating, newest, bestseller]
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: productId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /quiz/result:
    post:
      operationId: getQuizResult
      tags: [products]
      summary: Get recommended products based on quiz answers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizResultBody"
      responses:
        "200":
          description: Recommended products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Product:
      type: object
      required: [id, name, slug, price, categoryId, categorySlug, categoryName, inStock]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
        imageUrl:
          type: string
        categoryId:
          type: integer
        categorySlug:
          type: string
        categoryName:
          type: string
        strain:
          type: string
        thcContent:
          type: string
        cbdContent:
          type: string
        origin:
          type: string
        weight:
          type: string
        tags:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        isBestseller:
          type: boolean
        isNew:
          type: boolean
        inStock:
          type: boolean
        stock:
          type: integer
    CreateProductBody:
      type: object
      required: [name, price, categoryId]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        categoryId:
          type: integer
        imageUrl:
          type: string
        strain:
          type: string
        thcContent:
          type: string
        cbdContent:
          type: string
        tags:
          type: array
          items:
            type: string
        stock:
          type: integer
...
    Review:
      type: object
      required: [id, productId, authorName, rating, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        authorName:
          type: string
        authorLocation:
          type: string
        rating:
          type: integer
        body:
          type: string
        productVariant:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateReviewBody:
      type: object
      required: [productId, authorName, rating]
      properties:
        productId:
          type: integer
        authorName:
          type: string
        authorLocation:
          type: string
        rating:
          type: integer
        body:
          type: string
        productVariant:
          type: string
...
    QuizResultBody:
      type: object
      required: [mood]
      properties:
        mood:
          type: string
          enum: [serenity, energy, sleep, focus]
        goal:
          type: string
          nullable: true
          description: Raw answer to first quiz question
        experience:
          type: string
          nullable: true
          description: User CBD experience level
        format:
          type: string
          nullable: true
          description: Preferred consumption format